Create LasersPH
We currently use a ST-UT2 Tuned-Damped Upgradable Optical Table with automatic re-leveling
pneumatic vibration isolators from Newport. This line of optical tables can be upgraded to
include active damping if needed, although we have not found that to be necessary for our
experiments to date. We float the optical table using a house compressed air line.